+++ example/android/app/src/main/kotlin/com/mrrhak/khqr_sdk_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.mrrhak.khqr_sdk_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()