+++ lib/src/common/enum.dart
@@ -3,8 +3,13 @@
 /// - [KhqrCurrency.usd] Dollar (USD) - 840
 enum KhqrCurrency {
   /// Riel (KHR) - 116
-  khr,
+  khr(116),
 
   /// Dollar (USD) - 840
-  usd,
+  usd(840);
+
+  const KhqrCurrency(this.value);
+
+  /// The value of the enum
+  final int value;
 }